examples/a2a_adversarial_agent_example/stage2_improved/requirements.txt
# Stage 2: Improved Security Implementation Requirements
# Python 3.8 or higher required

# Authentication
PyJWT==2.8.0          # JWT token generation and verification
bcrypt==4.1.2         # Password hashing

# Optional for better development experience
# python-dotenv==1.0.0  # Environment variable management